AXON-707: restyle technical plan
