+++ src/react/atlascode/rovo-dev/rovoDevViewMessages.tsx
@@ -7,7 +7,6 @@ export const enum RovoDevViewResponseType {
     OpenFile = 'openFile',
     UndoFileChanges = 'undoFileChanges',
     KeepFileChanges = 'keepFileChanges',
-    GetOriginalText = 'getOriginalText',
     CreatePR = 'createPR',
     RetryPromptAfterError = 'retryPromptAfterError',
     GetCurrentBranchName = 'getCurrentBranchName',
@@ -29,7 +28,6 @@ export type RovoDevViewResponse =
     | ReducerAction<RovoDevViewResponseType.OpenFile, { filePath: string; tryShowDiff: boolean; range?: number[] }>
     | ReducerAction<RovoDevViewResponseType.UndoFileChanges, { files: ModifiedFile[] }>
     | ReducerAction<RovoDevViewResponseType.KeepFileChanges, { files: ModifiedFile[] }>
-    | ReducerAction<RovoDevViewResponseType.GetOriginalText, { filePath: string; range?: number[]; requestId: string }>
     | ReducerAction<RovoDevViewResponseType.CreatePR, { payload: { branchName: string; commitMessage: string } }>
     | ReducerAction<RovoDevViewResponseType.RetryPromptAfterError>
     | ReducerAction<RovoDevViewResponseType.GetCurrentBranchName>
