+++ src/rovo-dev/rovoDevWebviewProvider.ts
@@ -217,15 +217,6 @@ export class RovoDevWebviewProvider extends Disposable implements WebviewViewPro
                         await this.executeKeepFiles(e.files);
                         break;
 
-                    case RovoDevViewResponseType.GetOriginalText:
-                        const text = await this.executeGetText(e.filePath, e.range);
-                        await webviewView.webview.postMessage({
-                            type: RovoDevProviderMessageType.ReturnText,
-                            text: text || '',
-                            nonce: e.requestId, // Use the requestId as nonce
-                        });
-                        break;
-
                     case RovoDevViewResponseType.CreatePR:
                         await this.createPR(e.payload.commitMessage, e.payload.branchName);
                         break;
@@ -990,28 +981,6 @@ ${message}`;
         } else {
             await this.processError(new Error('RovoDev client not initialized'), false);
         }
-    }
-
-    private async executeGetText(filePath: string, range?: number[]): Promise<string | undefined> {
-        const resolvedPath = this.makeRelativePathAbsolute(filePath);
-        if (!fs.existsSync(resolvedPath)) {
-            console.warn(`File not found: ${resolvedPath}`);
-            return undefined;
-        }
-
-        const document = await workspace.openTextDocument(Uri.file(resolvedPath));
-
-        if (!document) {
-            console.warn(`Unable to open document for file: ${resolvedPath}`);
-            return undefined;
-        }
-
-        const lineRange =
-            range && Array.isArray(range) ? new Range(new Position(range[0], 0), new Position(range[1], 0)) : undefined;
-
-        const text = document.getText(document.validateRange(lineRange || new Range(0, 0, document.lineCount, 0)));
-
-        return text;
     }
 
     async invokeRovoDevAskCommand(prompt: string, context?: RovoDevContext): Promise<void> {
