+++ src/rovo-dev/rovoDevWebviewProviderMessages.ts
@@ -15,7 +15,6 @@ export const enum RovoDevProviderMessageType {
     NewSession = 'newSession',
     Initialized = 'initialized',
     CancelFailed = 'cancelFailed',
-    ReturnText = 'returnText',
     CreatePRComplete = 'createPRComplete',
     GetCurrentBranchNameComplete = 'getCurrentBranchNameComplete',
     UserFocusUpdated = 'userFocusUpdated',
@@ -37,7 +36,6 @@ export type RovoDevProviderMessage =
     | ReducerAction<RovoDevProviderMessageType.NewSession>
     | ReducerAction<RovoDevProviderMessageType.Initialized>
     | ReducerAction<RovoDevProviderMessageType.CancelFailed>
-    | ReducerAction<RovoDevProviderMessageType.ReturnText, { text: string }>
     | ReducerAction<RovoDevProviderMessageType.CreatePRComplete, { data: { url?: string; error?: string } }>
     | ReducerAction<RovoDevProviderMessageType.GetCurrentBranchNameComplete, { data: { branchName?: string } }>
     | ReducerAction<RovoDevProviderMessageType.UserFocusUpdated, { userFocus: RovoDevContextItem }>
